fix(#428): allow disabling authn entirely
