+++ client/src/core/context/authn/provider.tsx
@@ -1,8 +1,9 @@
+import { ErrorBoundary } from "core/context/error-boundary";
 import { createAuthnContext } from "core/context/initializers";
+import { Track } from "core/context/utils";
 import {
 	createContext,
 	onCleanup,
-	onMount,
 	Show,
 	type Accessor,
 	type Component,
@@ -13,20 +14,35 @@ import { useStore, type AuthnSvc } from "./store";
 export const AuthnContext =
 	createContext<Accessor<AuthnSvc>>(createAuthnContext);
 
-export const AuthnProvider: Component<ParentProps> = props => {
+const AuthnProviderWithoutErrorBoundary: Component<ParentProps> = props => {
 	const value = useStore();
 
-	onMount(() => {
-		value().actions.init();
-	});
+	const onMount = async () => {
+		return value().actions.init();
+	};
 
 	onCleanup(() => {
+		if (value().isInitialLoading) {
+			return;
+		}
+
 		value().actions.cleanup();
 	});
 
 	return (
-		<AuthnContext.Provider value={value}>
-			<Show when={!value().isInitialLoading}>{props.children}</Show>
-		</AuthnContext.Provider>
+		<>
+			<Track fn={onMount} />
+			<AuthnContext.Provider value={value}>
+				<Show when={!value().isInitialLoading}>{props.children}</Show>
+			</AuthnContext.Provider>
+		</>
 	);
 };
+
+export const AuthnProvider: Component<ParentProps> = props => (
+	<ErrorBoundary>
+		<AuthnProviderWithoutErrorBoundary>
+			{props.children}
+		</AuthnProviderWithoutErrorBoundary>
+	</ErrorBoundary>
+);
