+++ client/src/core/context/error-boundary/resources.ts
@@ -1,0 +1,9 @@
+export const resources = {
+	errorBoundary: {
+		title: "An unexpected error occurred",
+		description:
+			"Unable to initialize application. Please contact support.",
+		doOk: "Ok",
+		doCancel: "Cancel",
+	},
+};
